+++ R/accuracy.R
@@ -80,6 +80,20 @@ ACF1 <- function(.resid, na.action = stats::na.pass, demean = TRUE, ...){
 point_accuracy_measures <- list(ME = ME, RMSE = RMSE, MAE = MAE,
                        MPE = MPE, MAPE = MAPE, MASE = MASE, ACF1 = ACF1)
 
+#' Mean Arctangent Absolute Percentage Error
+#' 
+#' @inheritParams point_accuracy_measures
+#' 
+#' @references 
+#' Kim, Sungil and Heeyoung Kim (2016) "A new metric of absolute percentage error
+#' for intermittent demand forecasts". \emph{International Journal of Forecasting},
+#' \bold{32}(3), 669-679.
+#' 
+#' @export
+MAAPE <- function(.resid, .actual, na.rm = TRUE, ...){
+  mean(atan(abs(.resid / .actual * 100)), na.rm = na.rm)
+}
+
 #' @rdname interval_accuracy_measures
 #' @export
 winkler_score <- function(.dist, .actual, level = 95, na.rm = TRUE, ...){