+++ NEWS.md
@@ -9,6 +9,7 @@
 
 * Fixed `select.mdl_df()` usage with negative select values (#120).
 * Fixed `features()` for a tsibble with key variables but only one series.
+* Fixed interpolated values not being back transformed (tidyverts/fable#202).
 
 # fabletools 0.1.1