+++ NEWS.md
@@ -10,6 +10,7 @@
 * Fixed `select.mdl_df()` usage with negative select values (#120).
 * Fixed `features()` for a tsibble with key variables but only one series.
 * Fixed interpolated values not being back transformed (tidyverts/fable#202).
+* Fixed `stream()` causing issues with subsequent methods (#144).
 
 # fabletools 0.1.1